requirements.txt
cuquantum
cuquantum-python
cupy
matplotlib
numpy
qiskit-aer-gpu
ipykernel
pylatexenc
matplotlib